# contracts/generated/rust/openapi.json
{
  "components": {
    "schemas": {
...
      "ComponentStatus": {
        "enum": [
          "starting",
          "ready",
          "degraded",
          "failed"
        ],
        "type": "string"
      },
...
      "ReadinessSnapshot": {
        "properties": {
          "components": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ComponentStatus"
            },
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "status": {
            "$ref": "#/components/schemas/ReadinessStatus"
          }
        },
        "required": [
          "status",
          "components"
        ],
        "type": "object"
      },
      "ReadinessStatus": {
        "enum": [
          "starting",
          "ready",
          "degraded",
          "not_ready"
        ],
        "type": "string"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Music contributors"
    },
    "description": "",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "music-server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/readiness": {
      "get": {
        "operationId": "readiness",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessSnapshot"
                }
              }
            },
            "description": "All critical components can accept traffic"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessSnapshot"
                }
              }
            },
            "description": "A critical component is still starting or unavailable"
          }
        }
      }
    },
...
  }
}